[Refactor] Manual Install w install.php
Instead of copying in a LocalSettings.php from a wizard-assisted install
of MediaWiki, I create one automatically by doing a CLI installation in
a test-container. The test-container is built from the container I
intend to release on DockerHub.

This means that anyone pulling down the image on DockerHub will need to
run an install themselves and run the `maintenance/update.php` script
as well.